+++ packages/adguard-api-mv3/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - ???
+
+### Added
+
+- UserScripts API support.
+- New logger in internal libraries with tags.
+
+### Changed
+
+- Updated [@adguard/dnr-rulesets] to `v3.3.1`.
+
+[1.0.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/adguard-api-mv3-v1.0.0
+
 ## [0.4.0] - 2025-10-17
 
 ### Added